refactor: handle invalid pin labels in normalcomponent
